refactor(tools)!: restructure FirecrawlTool and update FirecrawlRequest
- Moved `FirecrawlTool` to `Tools\Scrape` namespace and implemented `ScrapeTool` interface.
- Simplified `FirecrawlRequest` by removing extraction prompt and schema definition.
- Adjusted dependencies and tests to reflect these changes. 🚀
