--- tests/Fixtures/Saloon/Tools/FirecrawlTool-4045f581b66dc3e16fa0edbf3869dc68.json
@@ -1,10 +1,0 @@
-{
-    "statusCode": 200,
-    "headers": {
-        "Date": "Fri, 04 Oct 2024 13:33:54 GMT",
-        "Content-Type": "application\/json",
-        "Transfer-Encoding": "chunked",
-        "Connection": "keep-alive"
-    },
-    "data": "{\n  \"id\": \"chatcmpl-AEcjJY2nt7KLnEhgSY83fSC7OXafR\",\n  \"object\": \"chat.completion\",\n  \"created\": 1728048833,\n  \"model\": \"gpt-4-turbo-2024-04-09\",\n  \"choices\": [\n    {\n      \"index\": 0,\n      \"message\": {\n        \"role\": \"assistant\",\n        \"content\": null,\n        \"tool_calls\": [\n          {\n            \"id\": \"call_sAg3DjOruYzDJrfEfJuk3AOT\",\n            \"type\": \"function\",\n            \"function\": {\n              \"name\": \"firecrawl_tool\",\n              \"arguments\": \"{\\\"url\\\":\\\"https:\/\/www.firecrawl.dev\/\\\",\\\"extractionPrompt\\\":\\\"Describe the main purpose or topic of the webpage.\\\"}\"\n            }\n          }\n        ],\n        \"refusal\": null\n      },\n      \"logprobs\": null,\n      \"finish_reason\": \"tool_calls\"\n    }\n  ],\n  \"usage\": {\n    \"prompt_tokens\": 141,\n    \"completion_tokens\": 36,\n    \"total_tokens\": 177,\n    \"prompt_tokens_details\": {\n      \"cached_tokens\": 0\n    },\n    \"completion_tokens_details\": {\n      \"reasoning_tokens\": 0\n    }\n  },\n  \"system_fingerprint\": \"fp_81dd8129df\"\n}\n"
-}
