+++ tests/Fixtures/Saloon/Tools/FirecrawlTool-76cdcfb02f44fe52a23d1c7248cd0cae.json
@@ -1,0 +1,14 @@
+{
+    "statusCode": 200,
+    "headers": {
+        "Date": "Tue, 22 Oct 2024 22:57:22 GMT",
+        "Content-Type": "application\/json",
+        "Content-Length": "1186",
+        "Connection": "keep-alive",
+        "access-control-expose-headers": "X-Request-ID",
+        "openai-organization": "user-5hvzt3x5aqwr0picqhxifwl8",
+        "openai-processing-ms": "3408",
+        "openai-version": "2020-10-01"
+    },
+    "data": "{\n  \"id\": \"chatcmpl-ALI6RfAUfMgMidYGCNGnNrynJLpTD\",\n  \"object\": \"chat.completion\",\n  \"created\": 1729637839,\n  \"model\": \"gpt-4-turbo-2024-04-09\",\n  \"choices\": [\n    {\n      \"index\": 0,\n      \"message\": {\n        \"role\": \"assistant\",\n        \"content\": \"```json\\n{\\n    \\\"answer\\\": \\\"The 'https:\/\/www.firecrawl.dev\/' page is about Firecrawl, a service that allows users to crawl and scrape data from websites, transforming it into clean, structured markdown or data that is ready for LLM (Large Language Models) applications. It provides various features including handling dynamic content, rotating proxies, and scraping without needing a sitemap. Firecrawl offers different pricing plans, is open source, and aims to power AI applications by providing reliable, clean data extracted from any website.\\\"\\n}\\n```\",\n        \"refusal\": null\n      },\n      \"logprobs\": null,\n      \"finish_reason\": \"stop\"\n    }\n  ],\n  \"usage\": {\n    \"prompt_tokens\": 5112,\n    \"completion_tokens\": 110,\n    \"total_tokens\": 5222,\n    \"prompt_tokens_details\": {\n      \"cached_tokens\": 0\n    },\n    \"completion_tokens_details\": {\n      \"reasoning_tokens\": 0\n    }\n  },\n  \"system_fingerprint\": \"fp_eae715ec6a\"\n}\n"
+}
